Finished full modularization and overall cleanup.
